requirements.txt
flask>=2.0.1
pytesseract>=0.3.8
Pillow
python-dotenv>=0.19.0
opencv-python
numpy
pandas
Werkzeug>=2.0.1